Se realizan cambios esteticos y se limpia codigo

+++ TP10_Angular/ng-proyecto-prueba/src/app/customers-table/customers-table.component.ts
@@ -12,10 +12,8 @@ import { Subscription, Observable } from 'rxjs';
 export class CustomersTableComponent implements OnInit {
   @Input() refreshCustomersTable: Boolean;
   @Output() sendOutParent = new EventEmitter<Customers>();
-  public customersData = new CustomersResponse();
   public customerData = new Customers();
   tableData: Observable<CustomersResponse>;
-  public subscription: Subscription;
   displayedColumns: string[] = ['ID', 'contactName', 'companyName', 'delete', 'update'];
 
   constructor(private northwindCustomersService: NorthwindCustomersService) {
@@ -32,17 +30,17 @@ export class CustomersTableComponent implements OnInit {
     }
   }
 
-  ngOnDestroy(): void {
-    this.subscription.unsubscribe;
-  }
 
   deleteCustomer(id) {
-    this.northwindCustomersService.deleteCustomers(id).subscribe(
-      resp => {
-        this.customerData = resp;
-        this.ngOnInit();
-      },
-      error => { alert(error.message); });
+    if (confirm("Esta seguro de querer borrar " + id + "?")) {
+      this.northwindCustomersService.deleteCustomers(id).subscribe(
+        resp => {
+          this.customerData = resp;
+          this.ngOnInit();
+        },
+        error => { alert(error.message); });
+
+    }
   }
 
   sendParent(customer: Customers) {
